Rename files to use capital CamelCase
