+++ shared/core/src/HousingCounselors.ts
