+++ shared/ui-components/src/page_components/HousingCounselor.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { HousingCounselor as Counselor } from "@bloom/core/src/housing_counselors"
+import { HousingCounselor as Counselor } from "@bloom/core/src/HousingCounselors"
 
 import t from "@bloom/ui-components/src/helpers/translator"
